A Single dialog for creating an Agent implemented.
Implemented DesignerEventListener's being called in NetworkDesigner.
Added getMaxConnections() to SocketAgent for retrieving the maximum
number of connections it is allowed to have.
The main Central Command application window now gets constructed and
initialized by the EDT thread.
Fixed AgentDebugWindow sending any messages with the message data set as
the recipient of the message.