Implemented Credentials' into the handshake
SocketAgent's can now be secured with a password to only allow
connections who have the correct credentials access. The protocol spec
talks about a bank of username and password pairs but this
implementation just deals with passwords but it wouldn't be too hard to
implement the one in the protocol spec. The code needs polishing but it
seems to work.