States refactored at the package level.
It was getting too messy to try to keep track of which states belonged
to the server, client or both. Client states are now in
denobo.connection.state.client and server states are in
denobo.connection.state.server. States for both (as well as the abstract
state) remain in denobo.connection.state.