Message dispatch cleanup implemented
Messages in the dispatch map whose route calculation are cleared are a
fixed interval.
The "Spam" button now causes less strain on the local system - the UI is
more responsive.
Added a "Stop Spam" button to the debug window.
The project should be considered feature complete as of now. Only
polishing, fixing and testing should be done in the remaining time.