Commenting, improvements and bug fixes
All javadoc and a few other warnings have been resolved.
Improved the ping refreshing in the monitor to run on a seperate thread
to prevent the UI locking up.
Fixed a NPE with the spam thread accessing Swing controls.
Fixed a threading issue with using RC4 encryption when an
RC4CryptoAlgorithm instance was been used to encrypt a packet at the
same time as it was decrypting one.